Implement first phase of two-phase conversion
This commit implements the first phase of the two-phase conversion, basically
everything unrelated to the actual output mode.

This also adds parsing of one more piece of information (paths to created
overlays -- currently an implementation detail which is not supposed to be
exposed and will be refactored later) into `log_parser.py`.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>