Run limited pylint tests
For now just run pylint that checks for errors (`-E`) because there are lot of
warnings from pylint's point of view, but instead of fixing all of them now, we
can gradually adopt more and more clean-ups to get to a better codebase.

Also add `pylint` to `requirements.txt` for developers and keep the
`tests_require` up to date even though it is not used (yet?).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>