Add _StateEncoder class
This class makes it possible to encode our own objects in the future.  In the
long run it should be possible to just `json.dumps(self)` in the state.  We can
make specialized classes for additional things in the state to make the code
more error-proof.

The exception is there due to a `pylint`'s false positive (already known
upstream and all over the internet).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>