log_parser: Fix JSON decoding error-catching
There is no `JSONDecodeError` on python2, so invalid JSON from `virt-v2v` could
just stop the whole migration.  This is not the intention, so catch `ValueError`
instead if `JSONDecodeError` does not exist as well.  Even though
`JSONDecodeError` inherits from ValueError and we can catch just that, I chose
to catch the most specific one, but I have no preference here.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>