Streamline error handling in the main function
This way we have one function (`finish()`) which is called *always*, no matter
what happens.  No need for double `try-except` and figuring out what to do based
on failure state in multiple places and so on.

Also account for the fact that some exceptions might not have arguments in which
case `e.args[0]` would throw an exception while handling an exception.

This commit makes mostly whitespace changes, so it is best viewed with `-w`.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>