Decrease some unnecessary nesting
This is not only for visual purposes, but also to keep pylint happy in future
commit which just moves code (and is hence easier to review with this change in
its own commit).  To be honest the change does not really make sense to me from
pylint's error (too-many-nested-blocks) point of view because in the future
commit where this becomes an issue there are less nested blocks, but in the end
I think this is easier to read anyway.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>