hosts: Add support for two_phase conversion with OpenstackHost
Add support for two_phase (and also transparently warm) conversion for the
openstack output.  Since virt-v2v does not do exactly what we would like it
to (yet) we need to introduce a wrapper for the openstack command and put it in
an additional `PATH` directory when running virt-v2v.  This command also has
tests for its usage (it should properly forward everything it does not
intercept).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>