ovirt: manage CR status better
There was an issue with infinte loop when the credentials were not
correct. In order to fix the issue we started to filter status
updates to CR. Now when there is an issue with credentials the
controller would attemt to reconnect for 5 minutes waiting for the
user to update the secret.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>