build: remove Kubevirt content from dist package
Do not include the Kubevirt content as it makes the package
unnecessarily large. The releases from master branch are the proper
source of the source files if anyone wishes to use the dist package (and
not to use git repo). As for the kubevirt v2v container the source is in
MangeIQ/manageiq-v2v-conversion_host-build repository.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>