wrapper: chain exceptions
When exception handler raises new exception Python 3 allows us to chain
the exceptions together. This provides more information in situation
when the final exception is not handled (which probably is not our case)
as it distinguishes the situation from the case when the error handler
fails for some other reason. And it also fixes linter error W0707.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>